refactor: Stop using deprecated base classes
