+++ src/mkdocstrings_handlers/python/__init__.py
@@ -3,3 +3,7 @@
 from mkdocstrings_handlers.python.handler import get_handler
 
 __all__ = ["get_handler"]  # noqa: WPS410
+
+# TODO: CSS classes everywhere in templates
+# TODO: name normalization (filenames, Jinja2 variables, HTML tags, CSS classes)
+# TODO: Jinja2 blocks everywhere in templates
